Unroll the settings checker under the comps upload instead of linking away
Somebody stands at the upload with a file in their hand and wonders whether
the settings in it were right. Sending them to another page to ask, and back
again to act on the answer, is three navigations for something that fits under
the form. The checker now opens in place, closed by default - an open drop
zone under an upload form is one more thing to mistake for the upload itself.

The page at /comps/check stays, rendering the same component. It is what a
refusal, a Discord answer or a link in chat can point at, for everybody who
arrives already asking.

Two paragraphs on the comps page were text-gray-600 on a dark panel: the
highlight explanation and the note about times staying hidden. Both had to be
selected with the mouse to be read, which is not a caption - it is a paragraph
that was published invisible.
