+++ lang/cs.json
@@ -4250,5 +4250,6 @@
     "before you run": "než pojedeš",
     "Not sure your settings were right?": "Nejsi si jistý nastavením?",
     "Check any demo": "Zkontroluj si jakékoliv demo",
-    "before you enter. Nothing is uploaded, and one second of recording is enough.": "než ho přihlásíš. Nic se nenahrává a stačí vteřina záznamu."
+    "before you enter. Nothing is uploaded, and one second of recording is enough.": "než ho přihlásíš. Nic se nenahrává a stačí vteřina záznamu.",
+    "Nothing is uploaded.": "Nic se nenahrává."
 }
