+++ lang/de.json
@@ -4250,5 +4250,6 @@
     "before you run": "bevor du läufst",
     "Not sure your settings were right?": "Unsicher, ob deine Einstellungen stimmten?",
     "Check any demo": "Prüfe ein beliebiges Demo",
-    "before you enter. Nothing is uploaded, and one second of recording is enough.": "bevor du teilnimmst. Nichts wird hochgeladen, und eine Sekunde Aufnahme reicht."
+    "before you enter. Nothing is uploaded, and one second of recording is enough.": "bevor du teilnimmst. Nichts wird hochgeladen, und eine Sekunde Aufnahme reicht.",
+    "Nothing is uploaded.": "Nichts wird hochgeladen."
 }
