+++ lang/es.json
@@ -4250,5 +4250,6 @@
     "before you run": "antes de correr",
     "Not sure your settings were right?": "¿No estás seguro de tus ajustes?",
     "Check any demo": "Comprueba cualquier demo",
-    "before you enter. Nothing is uploaded, and one second of recording is enough.": "antes de participar. No se sube nada y basta con un segundo de grabación."
+    "before you enter. Nothing is uploaded, and one second of recording is enough.": "antes de participar. No se sube nada y basta con un segundo de grabación.",
+    "Nothing is uploaded.": "No se sube nada."
 }
