+++ lang/fr.json
@@ -4250,5 +4250,6 @@
     "before you run": "avant de courir",
     "Not sure your settings were right?": "Vous doutez de vos réglages ?",
     "Check any demo": "Vérifiez n'importe quelle démo",
-    "before you enter. Nothing is uploaded, and one second of recording is enough.": "avant de participer. Rien n'est envoyé, et une seconde d'enregistrement suffit."
+    "before you enter. Nothing is uploaded, and one second of recording is enough.": "avant de participer. Rien n'est envoyé, et une seconde d'enregistrement suffit.",
+    "Nothing is uploaded.": "Rien n'est envoyé."
 }
