+++ lang/nl.json
@@ -4250,5 +4250,6 @@
     "before you run": "voordat je gaat rennen",
     "Not sure your settings were right?": "Niet zeker of je instellingen klopten?",
     "Check any demo": "Controleer elke demo",
-    "before you enter. Nothing is uploaded, and one second of recording is enough.": "voordat je meedoet. Er wordt niets geüpload en één seconde opnemen is genoeg."
+    "before you enter. Nothing is uploaded, and one second of recording is enough.": "voordat je meedoet. Er wordt niets geüpload en één seconde opnemen is genoeg.",
+    "Nothing is uploaded.": "Er wordt niets geüpload."
 }
