+++ lang/pl.json
@@ -4250,5 +4250,6 @@
     "before you run": "zanim pobiegniesz",
     "Not sure your settings were right?": "Nie masz pewności co do ustawień?",
     "Check any demo": "Sprawdź dowolne demo",
-    "before you enter. Nothing is uploaded, and one second of recording is enough.": "zanim je zgłosisz. Nic nie jest wysyłane, a sekunda nagrania wystarczy."
+    "before you enter. Nothing is uploaded, and one second of recording is enough.": "zanim je zgłosisz. Nic nie jest wysyłane, a sekunda nagrania wystarczy.",
+    "Nothing is uploaded.": "Nic nie jest wysyłane."
 }
