+++ lang/ru.json
@@ -4250,5 +4250,6 @@
     "before you run": "перед забегом",
     "Not sure your settings were right?": "Не уверены в своих настройках?",
     "Check any demo": "Проверьте любое демо",
-    "before you enter. Nothing is uploaded, and one second of recording is enough.": "перед участием. Ничего не загружается, и хватит одной секунды записи."
+    "before you enter. Nothing is uploaded, and one second of recording is enough.": "перед участием. Ничего не загружается, и хватит одной секунды записи.",
+    "Nothing is uploaded.": "Ничего не загружается."
 }
