+++ lang/sv.json
@@ -4250,5 +4250,6 @@
     "before you run": "innan du kör",
     "Not sure your settings were right?": "Osäker på dina inställningar?",
     "Check any demo": "Kontrollera vilken demo som helst",
-    "before you enter. Nothing is uploaded, and one second of recording is enough.": "innan du deltar. Ingenting laddas upp, och en sekunds inspelning räcker."
+    "before you enter. Nothing is uploaded, and one second of recording is enough.": "innan du deltar. Ingenting laddas upp, och en sekunds inspelning räcker.",
+    "Nothing is uploaded.": "Ingenting laddas upp."
 }
