+++ lang/uk.json
@@ -4250,5 +4250,6 @@
     "before you run": "перед забігом",
     "Not sure your settings were right?": "Не впевнені у своїх налаштуваннях?",
     "Check any demo": "Перевірте будь-яке демо",
-    "before you enter. Nothing is uploaded, and one second of recording is enough.": "перед участю. Нічого не завантажується, і вистачить однієї секунди запису."
+    "before you enter. Nothing is uploaded, and one second of recording is enough.": "перед участю. Нічого не завантажується, і вистачить однієї секунди запису.",
+    "Nothing is uploaded.": "Нічого не завантажується."
 }
